# Benchmark and a number of tuning steps for LFVM interpreter