--- a/core/vm/lfvm/interpreter.go
+++ b/core/vm/lfvm/interpreter.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/vm"
+	"github.com/ethereum/go-ethereum/params"
 	"github.com/holiman/uint256"
 )
 
@@ -412,11 +413,13 @@ func step(c *context) {
 		return
 	}
 
-	if c.stack.len() < staticStackBoundry[op].stackMin {
+	stackLen := c.stack.len()
+	if stackLen < staticStackBoundry[op].stackMin {
 		c.err = &vm.ErrStackUnderflow{}
 		c.status = ERROR
 		return
-	} else if c.stack.len() > staticStackBoundry[op].stackMax {
+	}
+	if stackLen > int(params.StackLimit)-1 && stackLen > staticStackBoundry[op].stackMax {
 		c.err = &vm.ErrStackOverflow{}
 		c.status = ERROR
 		return