--- a/core/vm/lfvm/interpreter.go
+++ b/core/vm/lfvm/interpreter.go
@@ -184,18 +184,7 @@ func Run(evm *vm.EVM, cfg vm.Config, contract *vm.Contract, code Code, data []by
 }
 
 func run(c *context) {
-	for c.status == RUNNING {
-		if int(c.pc) >= len(c.code) {
-			opStop(c)
-			return
-		}
-
-		// JUMP_TO is an LFVM specific operation that has no gas costs.
-		if c.code[c.pc].opcode == JUMP_TO {
-			c.pc = int32(c.code[c.pc].arg)
-		}
-		step(c)
-	}
+	stepToEnd(c)
 }
 
 func runWithShadowInterpreter(c *context) {
@@ -387,383 +376,406 @@ func runWithStatistics(c *context) {
 }
 
 func step(c *context) {
+	steps(c, true)
+}
 
-	op := c.code[c.pc].opcode
+func stepToEnd(c *context) {
+	steps(c, false)
+}
+func steps(c *context, one_step_only bool) {
+	for c.status == RUNNING {
+		if int(c.pc) >= len(c.code) {
+			opStop(c)
+			return
+		}
 
-	// Catch invalid op-codes here, to avoid the need to check them at other places multiple times.
-	if op >= NUM_EXECUTABLE_OPCODES {
-		c.err = vm.ErrInvalidCode
-		c.status = ERROR
-		return
-	}
+		op := c.code[c.pc].opcode
 
-	// If the interpreter is operating in readonly mode, make sure no
-	// state-modifying operation is performed. The 3rd stack item
-	// for a call operation is the value. Transferring value from one
-	// account to the others means the state is modified and should also
-	// return with an error.
-	if c.readOnly && (isWriteInstruction(op) || (op == CALL && c.stack.Back(2).Sign() != 0)) {
-		c.err = vm.ErrWriteProtection
-		c.status = ERROR
-		return
-	}
+		// JUMP_TO is an LFVM specific operation that has no gas costs.
+		if c.code[c.pc].opcode == JUMP_TO {
+			c.pc = int32(c.code[c.pc].arg)
+			op = c.code[c.pc].opcode
+		}
 
-	// Consume static gas price for instruction before execution
-	if !c.UseGas(getGasPrice(c, op)) {
-		return
-	}
+		// Catch invalid op-codes here, to avoid the need to check them at other places multiple times.
+		if op >= NUM_EXECUTABLE_OPCODES {
+			c.err = vm.ErrInvalidCode
+			c.status = ERROR
+			return
+		}
 
-	stackLen := c.stack.len()
-	if stackLen < staticStackBoundry[op].stackMin {
-		c.err = &vm.ErrStackUnderflow{}
-		c.status = ERROR
-		return
-	}
-	if stackLen > int(params.StackLimit)-1 && stackLen > staticStackBoundry[op].stackMax {
-		c.err = &vm.ErrStackOverflow{}
-		c.status = ERROR
-		return
-	}
+		// If the interpreter is operating in readonly mode, make sure no
+		// state-modifying operation is performed. The 3rd stack item
+		// for a call operation is the value. Transferring value from one
+		// account to the others means the state is modified and should also
+		// return with an error.
+		if c.readOnly && (isWriteInstruction(op) || (op == CALL && c.stack.Back(2).Sign() != 0)) {
+			c.err = vm.ErrWriteProtection
+			c.status = ERROR
+			return
+		}
 
-	// Execute instruction
-	switch op {
-	case POP:
-		opPop(c)
-	case PUSH1:
-		opPush1(c)
-	case PUSH2:
-		opPush2(c)
-	case PUSH3:
-		opPush3(c)
-	case PUSH4:
-		opPush4(c)
-	case PUSH5:
-		opPush(c, 5)
-	case PUSH31:
-		opPush(c, 31)
-	case PUSH32:
-		opPush32(c)
-	case JUMP:
-		opJump(c)
-	case JUMPDEST:
-		// nothing
-	case SWAP1:
-		opSwap(c, 1)
-	case SWAP2:
-		opSwap(c, 2)
-	case DUP3:
-		opDup(c, 3)
-	case AND:
-		opAnd(c)
-	case SWAP3:
-		opSwap(c, 3)
-	case JUMPI:
-		opJumpi(c)
-	case GT:
-		opGt(c)
-	case DUP4:
-		opDup(c, 4)
-	case DUP2:
-		opDup(c, 2)
-	case ISZERO:
-		opIszero(c)
-	case ADD:
-		opAdd(c)
-	case OR:
-		opOr(c)
-	case XOR:
-		opXor(c)
-	case NOT:
-		opNot(c)
-	case SUB:
-		opSub(c)
-	case MUL:
-		opMul(c)
-	case MULMOD:
-		opMulMod(c)
-	case DIV:
-		opDiv(c)
-	case SDIV:
-		opSDiv(c)
-	case MOD:
-		opMod(c)
-	case SMOD:
-		opSMod(c)
-	case ADDMOD:
-		opAddMod(c)
-	case EXP:
-		opExp(c)
-	case DUP5:
-		opDup(c, 5)
-	case DUP1:
-		opDup(c, 1)
-	case EQ:
-		opEq(c)
-	case PC:
-		opPc(c)
-	case CALLER:
-		opCaller(c)
-	case CALLDATALOAD:
-		opCallDataload(c)
-	case CALLDATASIZE:
-		opCallDatasize(c)
-	case CALLDATACOPY:
-		opCallDataCopy(c)
-	case MLOAD:
-		opMload(c)
-	case MSTORE:
-		opMstore(c)
-	case MSTORE8:
-		opMstore8(c)
-	case MSIZE:
-		opMsize(c)
-	case LT:
-		opLt(c)
-	case SLT:
-		opSlt(c)
-	case SGT:
-		opSgt(c)
-	case SHR:
-		opShr(c)
-	case SHL:
-		opShl(c)
-	case SAR:
-		opSar(c)
-	case SIGNEXTEND:
-		opSignExtend(c)
-	case BYTE:
-		opByte(c)
-	case SHA3:
-		opSha3(c)
-	case CALLVALUE:
-		opCallvalue(c)
-	case PUSH6:
-		opPush(c, 6)
-	case PUSH7:
-		opPush(c, 7)
-	case PUSH8:
-		opPush(c, 8)
-	case PUSH9:
-		opPush(c, 9)
-	case PUSH10:
-		opPush(c, 10)
-	case PUSH11:
-		opPush(c, 11)
-	case PUSH12:
-		opPush(c, 12)
-	case PUSH13:
-		opPush(c, 13)
-	case PUSH14:
-		opPush(c, 14)
-	case PUSH15:
-		opPush(c, 15)
-	case PUSH16:
-		opPush(c, 16)
-	case PUSH17:
-		opPush(c, 17)
-	case PUSH18:
-		opPush(c, 18)
-	case PUSH19:
-		opPush(c, 19)
-	case PUSH20:
-		opPush(c, 20)
-	case PUSH21:
-		opPush(c, 21)
-	case PUSH22:
-		opPush(c, 22)
-	case PUSH23:
-		opPush(c, 23)
-	case PUSH24:
-		opPush(c, 24)
-	case PUSH25:
-		opPush(c, 25)
-	case PUSH26:
-		opPush(c, 26)
-	case PUSH27:
-		opPush(c, 27)
-	case PUSH28:
-		opPush(c, 28)
-	case PUSH29:
-		opPush(c, 29)
-	case PUSH30:
-		opPush(c, 30)
-	case SWAP4:
-		opSwap(c, 4)
-	case SWAP5:
-		opSwap(c, 5)
-	case SWAP6:
-		opSwap(c, 6)
-	case SWAP7:
-		opSwap(c, 7)
-	case SWAP8:
-		opSwap(c, 8)
-	case SWAP9:
-		opSwap(c, 9)
-	case SWAP10:
-		opSwap(c, 10)
-	case SWAP11:
-		opSwap(c, 11)
-	case SWAP12:
-		opSwap(c, 12)
-	case SWAP13:
-		opSwap(c, 13)
-	case SWAP14:
-		opSwap(c, 14)
-	case SWAP15:
-		opSwap(c, 15)
-	case SWAP16:
-		opSwap(c, 16)
-	case DUP6:
-		opDup(c, 6)
-	case DUP7:
-		opDup(c, 7)
-	case DUP8:
-		opDup(c, 8)
-	case DUP9:
-		opDup(c, 9)
-	case DUP10:
-		opDup(c, 10)
-	case DUP11:
-		opDup(c, 11)
-	case DUP12:
-		opDup(c, 12)
-	case DUP13:
-		opDup(c, 13)
-	case DUP14:
-		opDup(c, 14)
-	case DUP15:
-		opDup(c, 15)
-	case DUP16:
-		opDup(c, 16)
-	case RETURN:
-		opReturn(c)
-	case REVERT:
-		opRevert(c)
-	case JUMP_TO:
-		opJumpTo(c)
-	case NOOP:
-		opNoop(c)
-	case DATA:
-		panic("can not interpret data as instruction")
-	case INVALID:
-		opInvalid(c)
-	case SLOAD:
-		opSload(c)
-	case SSTORE:
-		opSstore(c)
-	case CODESIZE:
-		opCodeSize(c)
-	case CODECOPY:
-		opCodeCopy(c)
-	case EXTCODESIZE:
-		opExtcodesize(c)
-	case EXTCODEHASH:
-		opExtcodehash(c)
-	case EXTCODECOPY:
-		opExtCodeCopy(c)
-	case BALANCE:
-		opBalance(c)
-	case SELFBALANCE:
-		opSelfbalance(c)
-	case BASEFEE:
-		opBaseFee(c)
-	case SELFDESTRUCT:
-		opSelfdestruct(c)
-	case CHAINID:
-		opChainId(c)
-	case GAS:
-		opGas(c)
-	case DIFFICULTY:
-		opDifficulty(c)
-	case TIMESTAMP:
-		opTimestamp(c)
-	case NUMBER:
-		opNumber(c)
-	case GASLIMIT:
-		opGasLimit(c)
-	case GASPRICE:
-		opGasPrice(c)
-	case CALL:
-		opCall(c)
-	case CALLCODE:
-		opCallCode(c)
-	case STATICCALL:
-		opStaticCall(c)
-	case DELEGATECALL:
-		opDelegateCall(c)
-	case RETURNDATASIZE:
-		opReturnDataSize(c)
-	case RETURNDATACOPY:
-		opReturnDataCopy(c)
-	case BLOCKHASH:
-		opBlockhash(c)
-	case COINBASE:
-		opCoinbase(c)
-	case ORIGIN:
-		opOrigin(c)
-	case ADDRESS:
-		opAddress(c)
-	case STOP:
-		opStop(c)
-	case CREATE:
-		opCreate(c)
-	case CREATE2:
-		opCreate2(c)
-	case LOG0:
-		opLog(c, 0)
-	case LOG1:
-		opLog(c, 1)
-	case LOG2:
-		opLog(c, 2)
-	case LOG3:
-		opLog(c, 3)
-	case LOG4:
-		opLog(c, 4)
-	// --- Super Instructions ---
-	case SWAP2_SWAP1_POP_JUMP:
-		opSwap2_Swap1_Pop_Jump(c)
-	case SWAP1_POP_SWAP2_SWAP1:
-		opSwap1_Pop_Swap2_Swap1(c)
-	case POP_SWAP2_SWAP1_POP:
-		opPop_Swap2_Swap1_Pop(c)
-	case POP_POP:
-		opPopPop(c)
-	case PUSH1_SHL:
-		opPush1_Shl(c)
-	case PUSH1_ADD:
-		opPush1_Add(c)
-	case PUSH1_DUP1:
-		opPush1_Dup1(c)
-	case PUSH2_JUMP:
-		opPush2_Jump(c)
-	case PUSH2_JUMPI:
-		opPush2_Jumpi(c)
-	case PUSH1_PUSH1:
-		opPush1_Push1(c)
-	case SWAP1_POP:
-		opSwap1_Pop(c)
-	case POP_JUMP:
-		opPop_Jump(c)
-	case SWAP2_SWAP1:
-		opSwap2_Swap1(c)
-	case SWAP2_POP:
-		opSwap2_Pop(c)
-	case DUP2_MSTORE:
-		opDup2_Mstore(c)
-	case DUP2_LT:
-		opDup2_Lt(c)
-	case ISZERO_PUSH2_JUMPI:
-		opIsZero_Push2_Jumpi(c)
-	case PUSH1_PUSH4_DUP3:
-		opPush1_Push4_Dup3(c)
-	case AND_SWAP1_POP_SWAP2_SWAP1:
-		opAnd_Swap1_Pop_Swap2_Swap1(c)
-	case PUSH1_PUSH1_PUSH1_SHL_SUB:
-		opPush1_Push1_Push1_Shl_Sub(c)
-	default:
-		panic(fmt.Sprintf("Unsupported operation: %v", op))
+		// Consume static gas price for instruction before execution
+		if !c.UseGas(getGasPrice(c, op)) {
+			return
+		}
+
+		stackLen := c.stack.len()
+		if stackLen < staticStackBoundry[op].stackMin {
+			c.err = &vm.ErrStackUnderflow{}
+			c.status = ERROR
+			return
+		}
+		if stackLen > int(params.StackLimit)-1 && stackLen > staticStackBoundry[op].stackMax {
+			c.err = &vm.ErrStackOverflow{}
+			c.status = ERROR
+			return
+		}
+
+		// Execute instruction
+		switch op {
+		case POP:
+			opPop(c)
+		case PUSH1:
+			opPush1(c)
+		case PUSH2:
+			opPush2(c)
+		case PUSH3:
+			opPush3(c)
+		case PUSH4:
+			opPush4(c)
+		case PUSH5:
+			opPush(c, 5)
+		case PUSH31:
+			opPush(c, 31)
+		case PUSH32:
+			opPush32(c)
+		case JUMP:
+			opJump(c)
+		case JUMPDEST:
+			// nothing
+		case SWAP1:
+			opSwap(c, 1)
+		case SWAP2:
+			opSwap(c, 2)
+		case DUP3:
+			opDup(c, 3)
+		case AND:
+			opAnd(c)
+		case SWAP3:
+			opSwap(c, 3)
+		case JUMPI:
+			opJumpi(c)
+		case GT:
+			opGt(c)
+		case DUP4:
+			opDup(c, 4)
+		case DUP2:
+			opDup(c, 2)
+		case ISZERO:
+			opIszero(c)
+		case ADD:
+			opAdd(c)
+		case OR:
+			opOr(c)
+		case XOR:
+			opXor(c)
+		case NOT:
+			opNot(c)
+		case SUB:
+			opSub(c)
+		case MUL:
+			opMul(c)
+		case MULMOD:
+			opMulMod(c)
+		case DIV:
+			opDiv(c)
+		case SDIV:
+			opSDiv(c)
+		case MOD:
+			opMod(c)
+		case SMOD:
+			opSMod(c)
+		case ADDMOD:
+			opAddMod(c)
+		case EXP:
+			opExp(c)
+		case DUP5:
+			opDup(c, 5)
+		case DUP1:
+			opDup(c, 1)
+		case EQ:
+			opEq(c)
+		case PC:
+			opPc(c)
+		case CALLER:
+			opCaller(c)
+		case CALLDATALOAD:
+			opCallDataload(c)
+		case CALLDATASIZE:
+			opCallDatasize(c)
+		case CALLDATACOPY:
+			opCallDataCopy(c)
+		case MLOAD:
+			opMload(c)
+		case MSTORE:
+			opMstore(c)
+		case MSTORE8:
+			opMstore8(c)
+		case MSIZE:
+			opMsize(c)
+		case LT:
+			opLt(c)
+		case SLT:
+			opSlt(c)
+		case SGT:
+			opSgt(c)
+		case SHR:
+			opShr(c)
+		case SHL:
+			opShl(c)
+		case SAR:
+			opSar(c)
+		case SIGNEXTEND:
+			opSignExtend(c)
+		case BYTE:
+			opByte(c)
+		case SHA3:
+			opSha3(c)
+		case CALLVALUE:
+			opCallvalue(c)
+		case PUSH6:
+			opPush(c, 6)
+		case PUSH7:
+			opPush(c, 7)
+		case PUSH8:
+			opPush(c, 8)
+		case PUSH9:
+			opPush(c, 9)
+		case PUSH10:
+			opPush(c, 10)
+		case PUSH11:
+			opPush(c, 11)
+		case PUSH12:
+			opPush(c, 12)
+		case PUSH13:
+			opPush(c, 13)
+		case PUSH14:
+			opPush(c, 14)
+		case PUSH15:
+			opPush(c, 15)
+		case PUSH16:
+			opPush(c, 16)
+		case PUSH17:
+			opPush(c, 17)
+		case PUSH18:
+			opPush(c, 18)
+		case PUSH19:
+			opPush(c, 19)
+		case PUSH20:
+			opPush(c, 20)
+		case PUSH21:
+			opPush(c, 21)
+		case PUSH22:
+			opPush(c, 22)
+		case PUSH23:
+			opPush(c, 23)
+		case PUSH24:
+			opPush(c, 24)
+		case PUSH25:
+			opPush(c, 25)
+		case PUSH26:
+			opPush(c, 26)
+		case PUSH27:
+			opPush(c, 27)
+		case PUSH28:
+			opPush(c, 28)
+		case PUSH29:
+			opPush(c, 29)
+		case PUSH30:
+			opPush(c, 30)
+		case SWAP4:
+			opSwap(c, 4)
+		case SWAP5:
+			opSwap(c, 5)
+		case SWAP6:
+			opSwap(c, 6)
+		case SWAP7:
+			opSwap(c, 7)
+		case SWAP8:
+			opSwap(c, 8)
+		case SWAP9:
+			opSwap(c, 9)
+		case SWAP10:
+			opSwap(c, 10)
+		case SWAP11:
+			opSwap(c, 11)
+		case SWAP12:
+			opSwap(c, 12)
+		case SWAP13:
+			opSwap(c, 13)
+		case SWAP14:
+			opSwap(c, 14)
+		case SWAP15:
+			opSwap(c, 15)
+		case SWAP16:
+			opSwap(c, 16)
+		case DUP6:
+			opDup(c, 6)
+		case DUP7:
+			opDup(c, 7)
+		case DUP8:
+			opDup(c, 8)
+		case DUP9:
+			opDup(c, 9)
+		case DUP10:
+			opDup(c, 10)
+		case DUP11:
+			opDup(c, 11)
+		case DUP12:
+			opDup(c, 12)
+		case DUP13:
+			opDup(c, 13)
+		case DUP14:
+			opDup(c, 14)
+		case DUP15:
+			opDup(c, 15)
+		case DUP16:
+			opDup(c, 16)
+		case RETURN:
+			opReturn(c)
+		case REVERT:
+			opRevert(c)
+		case JUMP_TO:
+			opJumpTo(c)
+		case NOOP:
+			opNoop(c)
+		case DATA:
+			panic("can not interpret data as instruction")
+		case INVALID:
+			opInvalid(c)
+		case SLOAD:
+			opSload(c)
+		case SSTORE:
+			opSstore(c)
+		case CODESIZE:
+			opCodeSize(c)
+		case CODECOPY:
+			opCodeCopy(c)
+		case EXTCODESIZE:
+			opExtcodesize(c)
+		case EXTCODEHASH:
+			opExtcodehash(c)
+		case EXTCODECOPY:
+			opExtCodeCopy(c)
+		case BALANCE:
+			opBalance(c)
+		case SELFBALANCE:
+			opSelfbalance(c)
+		case BASEFEE:
+			opBaseFee(c)
+		case SELFDESTRUCT:
+			opSelfdestruct(c)
+		case CHAINID:
+			opChainId(c)
+		case GAS:
+			opGas(c)
+		case DIFFICULTY:
+			opDifficulty(c)
+		case TIMESTAMP:
+			opTimestamp(c)
+		case NUMBER:
+			opNumber(c)
+		case GASLIMIT:
+			opGasLimit(c)
+		case GASPRICE:
+			opGasPrice(c)
+		case CALL:
+			opCall(c)
+		case CALLCODE:
+			opCallCode(c)
+		case STATICCALL:
+			opStaticCall(c)
+		case DELEGATECALL:
+			opDelegateCall(c)
+		case RETURNDATASIZE:
+			opReturnDataSize(c)
+		case RETURNDATACOPY:
+			opReturnDataCopy(c)
+		case BLOCKHASH:
+			opBlockhash(c)
+		case COINBASE:
+			opCoinbase(c)
+		case ORIGIN:
+			opOrigin(c)
+		case ADDRESS:
+			opAddress(c)
+		case STOP:
+			opStop(c)
+		case CREATE:
+			opCreate(c)
+		case CREATE2:
+			opCreate2(c)
+		case LOG0:
+			opLog(c, 0)
+		case LOG1:
+			opLog(c, 1)
+		case LOG2:
+			opLog(c, 2)
+		case LOG3:
+			opLog(c, 3)
+		case LOG4:
+			opLog(c, 4)
+		// --- Super Instructions ---
+		case SWAP2_SWAP1_POP_JUMP:
+			opSwap2_Swap1_Pop_Jump(c)
+		case SWAP1_POP_SWAP2_SWAP1:
+			opSwap1_Pop_Swap2_Swap1(c)
+		case POP_SWAP2_SWAP1_POP:
+			opPop_Swap2_Swap1_Pop(c)
+		case POP_POP:
+			opPopPop(c)
+		case PUSH1_SHL:
+			opPush1_Shl(c)
+		case PUSH1_ADD:
+			opPush1_Add(c)
+		case PUSH1_DUP1:
+			opPush1_Dup1(c)
+		case PUSH2_JUMP:
+			opPush2_Jump(c)
+		case PUSH2_JUMPI:
+			opPush2_Jumpi(c)
+		case PUSH1_PUSH1:
+			opPush1_Push1(c)
+		case SWAP1_POP:
+			opSwap1_Pop(c)
+		case POP_JUMP:
+			opPop_Jump(c)
+		case SWAP2_SWAP1:
+			opSwap2_Swap1(c)
+		case SWAP2_POP:
+			opSwap2_Pop(c)
+		case DUP2_MSTORE:
+			opDup2_Mstore(c)
+		case DUP2_LT:
+			opDup2_Lt(c)
+		case ISZERO_PUSH2_JUMPI:
+			opIsZero_Push2_Jumpi(c)
+		case PUSH1_PUSH4_DUP3:
+			opPush1_Push4_Dup3(c)
+		case AND_SWAP1_POP_SWAP2_SWAP1:
+			opAnd_Swap1_Pop_Swap2_Swap1(c)
+		case PUSH1_PUSH1_PUSH1_SHL_SUB:
+			opPush1_Push1_Push1_Shl_Sub(c)
+		default:
+			panic(fmt.Sprintf("Unsupported operation: %v", op))
+		}
+		c.pc++
+
+		if one_step_only {
+			return
+		}
 	}
-	c.pc++
 }
 
 func getGasPrice(c *context, op OpCode) uint64 {