--- a/core/vm/lfvm/interpreter_test.go
+++ b/core/vm/lfvm/interpreter_test.go
@@ -119,3 +119,11 @@ func BenchmarkFib10(b *testing.B) {
 func BenchmarkFib10_SI(b *testing.B) {
 	benchmarkFib(b, 10, true)
 }
+
+var sink bool
+
+func BenchmarkIsWriteInstruction(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		sink = isWriteInstruction(OpCode(i % int(NUM_EXECUTABLE_OPCODES)))
+	}
+}