This PR adds a benchmark for interpreting a program using the LFVM and applies a number of improvements to reduce the runtime of the interpreter (~ 35% reduction without super instructions and ~ 30-40% reduction with super instructions).

Before:
```
cpu: AMD Ryzen 9 5950X 16-Core Processor
BenchmarkFib10-32                  83112            142676 ns/op
BenchmarkFib10_SI-32              115978             99989 ns/op
cpu: Intel(R) Core(TM) i7-5820K CPU @ 3.30GHz
BenchmarkFib10-12       	   35463	    334699 ns/op
BenchmarkFib10_SI-12    	   54075	    235483 ns/op
```

After:
```
cpu: AMD Ryzen 9 5950X 16-Core Processor
BenchmarkFib10-32                 135072             92556 ns/op       
BenchmarkFib10_SI-32              168900             70866 ns/op 
cpu: Intel(R) Core(TM) i7-5820K CPU @ 3.30GHz
BenchmarkFib10-12       	   62058	    205308 ns/op
BenchmarkFib10_SI-12    	   90363	    134781 ns/op
```